requirements.txt
streamlit
pandas
sqlalchemy
pydantic
python-dotenv
apscheduler
requests
praw
rapidfuzz